Added ability to forward command
The intercepted command can now be passed to a forwarding command as an
argument, with a set of other arguments also passed to the forwarding
command.
The passed arguments can also be subject to variable replacement, where
the replacement values can be defined through Python snippets executed
by the 'eval()' funtion.
This allows things like the current path, or user id to be found and
passed in as arguments.